leo_model_package/requirements.txt
# Quick Draw CNN Training Requirements

# Deep Learning Framework
tensorflow>=2.10.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Visualization (optional, for plotting training curves)
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.62.0

# JSON handling (usually built-in, but just in case)
jsonlines>=2.0.0

# For saving and loading models
h5py>=3.1.0
